added consideration of reset period, must clean up

+++ RepeaterProtocol.py
@@ -1,9 +1,13 @@
+import numpy as np
+import netsquid as ns
 from netsquid.protocols import NodeProtocol, Signals
 from netsquid.components import QuantumProgram, instructions as instr
 
+from netsquid.components import Clock
+
 __all__ = [
     "RepeaterProtocol",
-    "RouteQubits",
+    "ManageRouting",
 ]
 
 class RepeaterProtocol(NodeProtocol):
@@ -23,9 +27,9 @@ class RepeaterProtocol(NodeProtocol):
 
     Subprotocols
     ------------
-    route_node_A : class:'RouteQubits'
+    route_node_A : class:'ManageRouting'
         Manages routing qubits incoming from node_A to available memory slot
-    route_node_B : class:'RouteQubits'
+    route_node_B : class:'ManageRouting'
         Manages routing qubits incoming from node_B to available memory slot
 
     """
@@ -33,17 +37,23 @@ class RepeaterProtocol(NodeProtocol):
         super().__init__(node, name=name)
         self.use_memory = use_memory
         self.result = None
-        self.qubits_A = []                                                                                              # memory positions of stored qubits from A->R channel
-        self.qubits_B = []                                                                                              # memory positions of stored qubits from B->R channel
         self._bsm_results = [(0, 0), (0, 1), (1, 0), (1, 1)]                                                            # Bell state measurement results
         self.port_names = list(dict(self.node.ports).keys())                                                            # Ports to end nodes generated in "setup_network"
         self._add_subprotocols(node,
                                self.port_names[0],
                                self.port_names[1])
 
+        self.slots_A = self.node.qmemory.get_matching_positions(
+            'origin', value='node_A')
+        self.slots_B = self.node.qmemory.get_matching_positions(
+            'origin', value='node_B')
+
+        self.qubits_A = dict((slot, False) for slot in self.slots_A)                                                    # FIXME: a bit weird implementation
+        self.qubits_B = dict((slot, False) for slot in self.slots_B)
+
     def _add_subprotocols(self, node, port_name_A, port_name_B):
-        self.add_subprotocol(RouteQubits(node, port_name_A, name='route_node_A'))
-        self.add_subprotocol(RouteQubits(node, port_name_B, name='route_node_B'))
+        self.add_subprotocol(ManageMemory(node, port_name_A, 'manage_slots_node_A'))
+        self.add_subprotocol(ManageMemory(node, port_name_B, 'manage_slots_node_B'))
 
     def run(self):
         if self.use_memory is False:
@@ -51,34 +61,37 @@ class RepeaterProtocol(NodeProtocol):
         else:
             self.start_subprotocols()
             while True:
-                expr = yield self.await_signal(self.subprotocols['route_node_A'], Signals.SUCCESS) | \
-                             self.await_signal(self.subprotocols['route_node_B'], Signals.SUCCESS)                      # wait for incoming qubit(s) to be successfully stored
-                if expr.first_term.value:                                                                               # update used memory slots with result from either signal
-                    mem_pos = self.subprotocols['route_node_A'].get_signal_result(
-                        label=Signals.SUCCESS, receiver=self)
-                    self.qubits_A.append(mem_pos)
-                if expr.second_term.value:
-                    mem_pos = self.subprotocols['route_node_B'].get_signal_result(
-                        label=Signals.SUCCESS, receiver=self)
-                    self.qubits_B.append(mem_pos)
+                expr = yield self.await_signal(self.subprotocols['manage_slots_node_A'], "OPERATION") | \
+                             self.await_signal(self.subprotocols['manage_slots_node_B'], "OPERATION")
 
-                while (len(self.qubits_A) > 0) and (len(self.qubits_B) > 0):                                            # both qubits from node_A and node_B are available in qmemory
-                    result = {
-                         'slots': [self.qubits_A.pop(0), self.qubits_B.pop(0)]                                            # FIXME: FIFO vs FILO
-                    }
-                    self.send_signal(Signals.SUCCESS, result=result)
-                    # measure_program = BellMeasurementProgram()
-                    # self.node.qmemory.execute_program(measure_program, [self.qubits_A[-1], self.qubits_B[-1]])
-                    # yield self.await_program(self.node.qmemory)
-                    #
-                    # m, = measure_program.output["BellStateIndex"]
-                    # m0, m1 = self._bsm_results[m]
-                    # measurement = [m0, m1]
-                    # result = {
-                    #     "meas": measurement,
-                    # #     "slots": [self.qubits_A.pop(), self.qubits_B.pop()]
-                    # # }
-                    # self.send_signal(Signals.SUCCESS,result=result)                                                   # FIXME: Could this cause issues if keeping track of elapsed time is important?
+                # manage instructions for memory A
+                if expr.first_term.value:
+                    operation = self.subprotocols['manage_slots_node_A'].get_signal_result("OPERATION", receiver=self)
+                    if operation["op"] == "ADD":
+                        self.qubits_A[operation["slot"]] = True
+                    elif operation["op"] == "REMOVE":
+                        self.qubits_A[operation["slot"]] = False
+
+                # manage instruction for memory B
+                else:
+                    operation = self.subprotocols['manage_slots_node_B'].get_signal_result("OPERATION", receiver=self)
+                    if operation["op"] == "ADD":
+                        self.qubits_B[operation["slot"]] = True
+                    elif operation["op"] == "REMOVE":
+                        self.qubits_B[operation["slot"]] = False
+
+                # FIXME: MUST BE BETTER WAY TO DO THIS PAIR  CHECKING
+                for idx_A in self.slots_A:
+                    if self.qubits_A[idx_A] is True:
+                        for idx_B in self.slots_B:
+                            if self.qubits_B[idx_B] is True:
+                                result = {
+                                     'slots': [idx_A, idx_B]
+                                }
+                                self.qubits_A[idx_A] = False
+                                self.qubits_B[idx_B] = False
+                                self.send_signal(Signals.SUCCESS, result=result)
+
 
     def _run_no_mem(self):                                                                                              # special run case for when memory should not store qubit for any measurable time
         while True:
@@ -96,20 +109,78 @@ class RepeaterProtocol(NodeProtocol):
                         'slots': [0, 1]
                     }
                     self.send_signal(Signals.SUCCESS, result=result)
-                # measure_program = BellMeasurementProgram()
-                # self.node.qmemory.execute_program(measure_program, [0, 1])
-                # yield self.await_program(self.node.qmemory)
-                # m, = measure_program.output["BellStateIndex"]
-                # m0, m1 = self._bsm_results[m]
-                # measurement = [m0, m1]
-                # result = {
-                #     "meas": measurement,
-                #     "slots": [0,1]                                                                                    # "no_mem" always accessed slots [0,1] for use with qprocessor (non-physical to avoid qmemory error)
-                # }
-                # self.send_signal(Signals.SUCCESS, result=result)  # FIXME: Could this cause issues if keeping track of elapsed time is important?
 
 
-class RouteQubits(NodeProtocol):
+class ManageMemory(NodeProtocol):
+    def __init__(self, node, port_name, name):
+        super().__init__(node, name=name)
+        self.operation = "OPERATION"
+        self.add_signal(self.operation)
+        # determine which slots will be managed
+        self.port_name = port_name
+        self.slots = self.node.qmemory.get_matching_positions(
+            'origin', value='{}'.format(self.name.replace('manage_slots_','')))
+        self._add_subprotocols(node)
+
+    def _add_subprotocols(self, node):
+        self.add_subprotocol(ManageRouting(node, self.port_name, self.slots, name='route'))
+        self.add_subprotocol(MemoryAccess(node, self.slots, name='access'))
+
+    def run(self):
+        self.start_subprotocols()
+        while True:
+            expr = yield self.await_signal(self.subprotocols['route'], "ADDED") | \
+                         self.await_signal(self.subprotocols['access'], "REMOVED")
+
+            if expr.first_term.value:
+                slot = self.subprotocols['route'].get_signal_result(label="ADDED", receiver=self)
+                operation = {"op": "ADD", "slot": slot}
+            else:
+                slot = self.subprotocols['access'].get_signal_result(label="REMOVED", receiver=self)
+                operation = {"op": "REMOVE", "slot": slot}
+
+
+            self.send_signal(self.operation, operation)
+
+
+class MemoryAccess(NodeProtocol):   # FIXME: should start once qubits are expected to arrive, using delay and max ticks
+
+    def __init__(self, node, slots, name):
+        super().__init__(node, name=name)
+        self.clock = Clock('internal', frequency=100e6, start_delay=166756.66666666666, max_ticks=10000)                               # FIXME: ALL HARDCODED
+        self.slots = slots
+        self.slots_status = [True] * len(self.slots)
+        self.reset_time = 300
+        self.slots_reset_timer = [self.reset_time] * len(self.slots)
+
+        self.removed_qubit = "REMOVED"
+        self.add_signal(self.removed_qubit)
+
+    def run(self):
+        self.clock.start()
+        while True:
+            yield self.await_port_output(self.clock.ports['cout'])
+
+            for idx, count in enumerate(self.slots_reset_timer):
+                if self.slots_status[idx] is False:
+                    self.slots_reset_timer[idx] -= 1
+                    if self.slots_reset_timer[idx] == 0:
+                        self.slots_status[idx] = True
+                        self.slots_reset_timer[idx] = self.reset_time
+                        self.node.qmemory.mem_positions[self.slots[idx]].reset()
+                        self.node.qmemory.set_position_used(False, idx)
+
+            # dont use "status == true" , instead only see if countdown is reset
+            if self.clock.num_ticks % 50 == 0:
+                for idx, status in enumerate(self.slots_status):
+                    if status is True:
+                        self.slots_status[idx] = False
+                        self.node.qmemory.set_position_used(True, idx)
+                        self.send_signal(self.removed_qubit, self.slots[idx])
+                        break
+
+
+class ManageRouting(NodeProtocol):
     """Subprotocol of "RepeaterProtocol": manages routing incoming qubits
 
     * Determines location of available slot (if any)
@@ -126,27 +197,28 @@ class RouteQubits(NodeProtocol):
         specify name of subprotocol
 
     """
-    def __init__(self, node, port_name, name):
+    def __init__(self, node, port_name, slots, name):
         super().__init__(node, name=name)
         self.port = self.node.ports[port_name]
+        self.slots = slots
+        self.new_qubit = "ADDED"
+        self.add_signal(self.new_qubit)
 
     def _get_unused_slot(self):                                                                                    # check for first available memory slot
-        for position in self.node.qmemory.get_matching_positions(
-                'origin', value='{}'.format(self.name.strip('route_'))):
-
-            if not self.node.qmemory.get_position_used(position):
-                return position
+        for slot in self.slots:
+            if not self.node.qmemory.get_position_used(slot):
+                return slot
         return None
 
     def run(self):
         while True:
-            yield self.await_port_input(self.port)                                                                      # wait for incoming qubit
+            yield self.await_port_input(self.port)                                                           # wait for incoming qubit
             message = self.port.rx_input()
             qubit, = message.items
             mem_pos = self._get_unused_slot()
             if mem_pos is not None:
                 self.node.qmemory.put(qubit, positions=mem_pos)
-                self.send_signal(Signals.SUCCESS, result=mem_pos)
+                self.send_signal(self.new_qubit, result=mem_pos)
 
 
 class BellMeasurementProgram(QuantumProgram):
