+++ SimulationProtocol.py
@@ -1,3 +1,4 @@
+import netsquid as ns
 from netsquid.protocols import LocalProtocol, Signals
 from netsquid.qubits import qubitapi as qapi, ketstates as ks
 from RepeaterProtocol import RepeaterProtocol
@@ -47,17 +48,20 @@ class SimulationProtocol(LocalProtocol):
         self.start_subprotocols()
         while True:
             yield self.await_signal(self.subprotocols['repeater_R'], Signals.SUCCESS)
+
             repeater_result = self.subprotocols['repeater_R'].get_signal_result(label=Signals.SUCCESS, receiver=self)
             slots = repeater_result['slots']
+            # FIXME: ISSUE WITH q1, q2 coming in as None (likely due to being removed by mem access protocol (SHOULD THIS CONDITIONAL BE NEEDED?)
             q1,q2 = self.subprotocols['repeater_R'].node.qmemory.pop(slots)                                             # grab/remove qubits after measurement FIXME: Figure out actual interpretation of results
-            fid_q1 = qapi.fidelity(q1, ks.y0, squared=True)
-            fid_q2 = qapi.fidelity(q2, ks.y0, squared=True)
-            fid_joint = qapi.fidelity([q1,q2], ks.y00, squared=True)
-            result = {
-                'fid_q1': fid_q1,
-                'pos_A': slots[0],
-                'fid_q2': fid_q2,
-                'pos_B': slots[1],
-                'fid_joint': fid_joint
-            }
-            self.send_signal(Signals.SUCCESS, result=result)
+            if q1 is not None and q2 is not None:
+                fid_q1 = qapi.fidelity(q1, ks.y0, squared=True)
+                fid_q2 = qapi.fidelity(q2, ks.y0, squared=True)
+                fid_joint = qapi.fidelity([q1,q2], ks.y00, squared=True)
+                result = {
+                    'fid_q1': fid_q1,
+                    'pos_A': slots[0],
+                    'fid_q2': fid_q2,
+                    'pos_B': slots[1],
+                    'fid_joint': fid_joint
+                }
+                self.send_signal(Signals.SUCCESS, result=result)
